refactor: NetworkManagerable, LocationManagerable을 별도 파일로 분리
