+++ WeatherForecast/WeatherForecast/Utilities/Implementation/LocationManager.swift
@@ -1,16 +1,6 @@
 import CoreLocation
 import UIKit
 
-typealias LocationCompletion = (Result<(CLLocationCoordinate2D, CLPlacemark), LocationError>) -> Void
-
-protocol LocationManagerable {
-    func request(coordinate: CLLocationCoordinate2D?, completion: @escaping LocationCompletion)
-}
-
-extension LocationManagerable {
-    func request(coordinate: CLLocationCoordinate2D?, completion: @escaping LocationCompletion) { }
-}
-
 final class LocationManager: NSObject, LocationManagerable {
     private let manager = CLLocationManager()
     private var locationCompletion: LocationCompletion?
